Remove FAQs for conciseness
GBT is now a legacy driver. Not everything in the FAQs was about GBT, and what was written about GB Studio in those FAQs is now outdated. Notably, not having SFX, "corrupted" .mod files, and requiring F96 on fast songs. What used to be in the FAQs could serve the community better in version-specific tutorials (videos, blog articles) rather than legacy user documentation.